apps/sources/urls.py
from rest_framework.routers import DefaultRouter

from apps.sources.views import SourceViewSet

router = DefaultRouter()
router.register(r"", SourceViewSet, basename="source")

urlpatterns = router.urls
