requirements/development.txt
-r base.txt

# Testing
pytest>=8.0
pytest-django>=4.8
pytest-asyncio>=0.23
factory-boy>=3.3
coverage>=7.4
pytest-cov>=5.0

# Dev tools
django-debug-toolbar>=4.3
ipython>=8.0
black>=24.0
isort>=5.13
flake8>=7.0

# Type checking
mypy>=1.9
django-stubs>=5.0
djangorestframework-stubs>=3.15
